+++ CHANGELOG.md
@@ -28,8 +28,12 @@ And constructed with the following guidelines:
 
 ***************************************************************************************************************
 
-## Version 0.0.5 | 2018-02-26
+## Version 0.0.5 | 2018-03-12
 
+* linted files
+* add unified school list for dropdown selection
+* added highlight of selected school to AzMERIT scatter chart
+* updated html for better responsive design
 * updated Kendo UI v2018.1.221 `UI R1 2018 SP1` [UI R1 2018 SP1](https://www.telerik.com/support/whats-new/kendo-ui/release-history/kendo-ui-r1-2018-sp1)
 * updated to ArcGIS API for JavaScript `v3.23` - [v3.23](https://developers.arcgis.com/javascript/)